requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
executing==2.2.1
Flask==3.1.2
fonttools==4.61.1
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.1.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
langchain-core==1.2.8
langchain-openai==1.1.7
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.8
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mmh3==4.1.0
nest-asyncio==1.6.0
nltk==3.9.2
numpy==2.4.2
openai==2.16.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.0
parso==0.8.5
pillow==12.1.0
pinecone-client==6.0.0
pinecone-notebooks==0.1.1
pinecone-plugin-interface==0.0.7
pinecone-text==0.11.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
stack-data==0.6.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
wcwidth==0.5.3
Werkzeug==3.1.5
xgboost==3.1.3
xxhash==3.6.0
zstandard==0.25.0
